put in package and remove javadocs
