+++ src/main/java/cc/polyfrost/oneconfig/config/core/exceptions/InvalidTypeException.java
@@ -1,0 +1,26 @@
+package cc.polyfrost.oneconfig.config.core.exceptions;
+
+public class InvalidTypeException extends RuntimeException {
+
+    public InvalidTypeException() {
+        super();
+    }
+
+    public InvalidTypeException(String message) {
+        super(message);
+    }
+
+    public InvalidTypeException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public InvalidTypeException(Throwable cause) {
+        super(cause);
+    }
+
+    protected InvalidTypeException(String message, Throwable cause,
+                                   boolean enableSuppression,
+                                   boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
